requirements.txt
aws-cdk-lib>=2.0.0rc1
constructs>=10.0.0
